requirements.txt
Django>=3.2
django-ninja